+++ tests/test_pytest_mypy.py
@@ -22,7 +22,7 @@ def test_mypy_error(testdir):
     result = testdir.runpytest_subprocess('--mypy')
     result.assert_outcomes(failed=1)
     result.stdout.fnmatch_lines([
-        'test_mypy_error.py:2: error: Incompatible return value*',
+        '2: error: Incompatible return value*',
     ])
     assert result.ret != 0
 
@@ -38,7 +38,7 @@ def test_mypy_ignore_missings_imports(testdir):
     result = testdir.runpytest_subprocess('--mypy')
     result.assert_outcomes(failed=1)
     result.stdout.fnmatch_lines([
-        '*1: error: Cannot find module named*',
+        '1: error: Cannot find module named*',
     ])
     assert result.ret != 0
     result = testdir.runpytest_subprocess('--mypy-ignore-missing-imports')
@@ -77,3 +77,44 @@ def test_non_mypy_error(testdir):
     result.assert_outcomes(failed=1)
     result.stdout.fnmatch_lines(['*' + message])
     assert result.ret != 0
+
+
+def test_mypy_stderr(testdir):
+    """Verify that stderr from mypy is printed."""
+    stderr = 'This is stderr from mypy.'
+    testdir.makepyfile(conftest='''
+        import mypy.api
+
+        def _patched_run(*args, **kwargs):
+            return '', '{stderr}', 1
+
+        mypy.api.run = _patched_run
+    '''.format(stderr=stderr))
+    result = testdir.runpytest_subprocess('--mypy')
+    result.stdout.fnmatch_lines([stderr])
+
+
+def test_mypy_unmatched_stdout(testdir):
+    """Verify that unexpected output on stdout from mypy is printed."""
+    stdout = 'This is unexpected output on stdout from mypy.'
+    testdir.makepyfile(conftest='''
+        import mypy.api
+
+        def _patched_run(*args, **kwargs):
+            return '{stdout}', '', 1
+
+        mypy.api.run = _patched_run
+    '''.format(stdout=stdout))
+    result = testdir.runpytest_subprocess('--mypy')
+    result.stdout.fnmatch_lines([stdout])
+
+
+def test_api_mypy_argv(testdir):
+    """Ensure that the plugin can be configured in a conftest.py."""
+    testdir.makepyfile(conftest='''
+        def pytest_configure(config):
+            plugin = config.pluginmanager.getplugin('mypy')
+            plugin.mypy_argv.append('--version')
+    ''')
+    result = testdir.runpytest_subprocess('--mypy')
+    assert result.ret == 0